requirements.txt
pyglm==2.8.3
pygame==2.6.1
moderngl==5.12.0
numpy==2.4.1